Move RunModel.py script into its own file
So it's more readable
